글, 계정삭제전에 한번 물어봄

+++ Login/static/js/detail.js
@@ -5,12 +5,12 @@ let Author = ""
 $(document).ready(function () {
     if (location.href.includes("?")) {
         postid = location.href.split("?")[1].split("postid=")[1]
-        if(postid.includes(";")) {
+        if (postid.includes(";")) {
             postid = postid.split(";")[0]
-            if(location.href.includes("post_author")) {
+            if (location.href.includes("post_author")) {
                 visitor = true
-                Author = location.href.split("post_author=")[1].replace(" ","")
-                if(Author.includes(";")) {
+                Author = location.href.split("post_author=")[1].replace(" ", "")
+                if (Author.includes(";")) {
                     Author = Author.split(";")[0]
                 }
                 let page = ""
@@ -54,22 +54,22 @@ $(document).ready(function () {
                             success: function (response) {
                                 response.forEach(element => {
                                     let now = new Date()
-        
+
                                     let year = now.getFullYear()
                                     let month = now.getMonth()
                                     let day = now.getDate()
                                     let hours = now.getHours() + 9
                                     let minutes = now.getMinutes()
                                     let seconds = now.getSeconds()
-        
+
                                     let nowdate = new Date(year, month, day, hours, minutes, seconds)
-        
+
                                     let chai = nowdate.getTime() - new Date(element.Created_At).getTime()
                                     let a = ""
-        
+
                                     let wmon = element.Created_At.split("-")[1]
                                     let wday = element.Created_At.split("-")[2].split("T")[0]
-        
+
                                     if (chai < 1000 * 60)
                                         a += '방금'
                                     else if (chai < 1000 * 60 * 60)
@@ -82,8 +82,8 @@ $(document).ready(function () {
                                         a += Math.floor(chai / (1000 * 60 * 60 * 24 * 30)) + '달 전';
                                     else
                                         a += Math.floor(chai / (1000 * 60 * 60 * 24 * 30 * 12)) + '년 전';
-        
-        
+
+
                                     if (postid == element.Created_At) {
                                         addCard(element)
                                     }
@@ -94,7 +94,7 @@ $(document).ready(function () {
                 })
             }
         }
-        
+
     } else {
         location.href = `/`;
         alert("표시할 글의 아이디가 지정되어있지 않습니다.\n카드에서 다시 클릭해보세요.\n홈으로 이동합니다.")
@@ -106,7 +106,7 @@ $(document).ready(function () {
         } else {
             id = document.cookie.split("user_id=")[1].replace(" ", "")
         }
-        
+
         let page = ""
         if (location.href.includes("?")) {
             let ihref = location.href.split("?")[1]
@@ -229,7 +229,7 @@ function addCard(element) {
 }
 
 function copyURL() {
-    navigator.clipboard.writeText(window.location.href+";post_author="+id);
+    navigator.clipboard.writeText(window.location.href + ";post_author=" + id);
     alert("클립보드에 링크를 복사했습니다!")
 }
 
@@ -251,8 +251,8 @@ function kakaoTalkShare() {
             {
                 title: '일기 보러가기',
                 link: {
-                    mobileWebUrl: window.location.href+";post_author="+id,
-                    webUrl: window.location.href+";post_author="+id,
+                    mobileWebUrl: window.location.href + ";post_author=" + id,
+                    webUrl: window.location.href + ";post_author=" + id,
                 },
             },
             {
@@ -268,35 +268,37 @@ function kakaoTalkShare() {
 
 function kakaoStoryShare() {
     Kakao.Story.share({
-        url: window.location.href+";post_author="+id,
+        url: window.location.href + ";post_author=" + id,
         text: '저의 일기를 공유합니다!'
     })
 }
 
 function twiterShare() {
-    const pageUrl = window.location.href+";post_author="+id
-    window.open(`https://twitter.com/intent/tweet?text=저의 일기를 공유합니다.\n&url=${pageUrl}`, '트위터 공유' , "width=870, height=880, resizable = no, scrollbars = no")
+    const pageUrl = window.location.href + ";post_author=" + id
+    window.open(`https://twitter.com/intent/tweet?text=저의 일기를 공유합니다.\n&url=${pageUrl}`, '트위터 공유', "width=870, height=880, resizable = no, scrollbars = no")
 }
 
 function facebookShare() {
-    const pageUrl = window.location.href+";post_author="+id
+    const pageUrl = window.location.href + ";post_author=" + id
     window.open(`http://www.facebook.com/sharer/sharer.php?u=${pageUrl}`, '페이스북 공유', "width=870, height=880, resizable = no, scrollbars = no")
 }
 
 function Delete() {
-    let created_At = location.href.split("?")[1].split("postid=")[1]
-    $.ajax({
-        url: '/Write',
-        method: "DELETE",
-        data: {
-            'id': id,
-            'Created_At': created_At
-        },
-        success: function (response) {
-            location.href = "/"
-            alert("글을 삭제했습니다.")
-        }
-    })
+    if (confirm("정말 삭제하시겠습니까?\n이 작업은 취소할수 없습니다!") == true) {
+        let created_At = location.href.split("?")[1].split("postid=")[1]
+        $.ajax({
+            url: '/Write',
+            method: "DELETE",
+            data: {
+                'id': id,
+                'Created_At': created_At
+            },
+            success: function (response) {
+                location.href = "/"
+                alert("글을 삭제했습니다.")
+            }
+        })
+    }
 }
 
 function CardI(name, adr, people_meet, image, date, feel, desc, Eat) {
@@ -305,7 +307,7 @@ function CardI(name, adr, people_meet, image, date, feel, desc, Eat) {
 
     }
 
-    if (visitor == false){
+    if (visitor == false) {
         return `<section id="section">
                     <div class="detail_wrap">
                         <div class="detail_section">
@@ -346,7 +348,7 @@ function CardI(name, adr, people_meet, image, date, feel, desc, Eat) {
                         </div>
                     </div>
                 </section>`
-    }else{
+    } else {
         return `<section id="section">
                     <div class="detail_wrap">
                         <div class="detail_section">
@@ -374,7 +376,7 @@ function CardI(name, adr, people_meet, image, date, feel, desc, Eat) {
                             </div>
                         </div>
                     </div>
-                </section>`        
+                </section>`
     }
 }
 function Card(name, adr, people_meet, date, feel, desc, Eat) {
@@ -421,7 +423,7 @@ function Card(name, adr, people_meet, date, feel, desc, Eat) {
                         </div>
                     </div>
                 </section>`
-    }else{
+    } else {
         return `<section id="section">
             <div class="detail_wrap">
                 <div class="detail_section">
